+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Make TODOs repo-wide, not per-file todos
 - Improve TODO screen and files screen TUI/TUX including hotkeys
-- Implemant branch rebase
+- Implemant branch rebase, drop
 - Fix TODOs autosave bug with todo duplication
 
 ## [0.1.6][] - 2026-09-15